+++ src/backend/query_builder.rs
@@ -255,9 +255,9 @@ pub trait QueryBuilder: QuotedBuilder + EscapeBuilder + TableRefBuilder {
                 write!(sql, " ").unwrap();
                 self.prepare_simple_expr(expr, sql);
             }
-            SimpleExpr::FunctionCall(func, exprs) => {
-                self.prepare_function(func, sql);
-                self.prepare_tuple(exprs, sql);
+            SimpleExpr::FunctionCall(func) => {
+                self.prepare_function(&func.func, sql);
+                self.prepare_tuple(&func.args, sql);
             }
             SimpleExpr::Binary(left, op, right) => {
                 if *op == BinOper::In && right.is_values() && right.get_values().is_empty() {
@@ -465,6 +465,12 @@ pub trait QueryBuilder: QuotedBuilder + EscapeBuilder + TableRefBuilder {
                 write!(sql, "(").unwrap();
                 self.prepare_values_list(values, sql);
                 write!(sql, ")").unwrap();
+                write!(sql, " AS ").unwrap();
+                alias.prepare(sql.as_writer(), self.quote());
+            }
+            TableRef::FunctionCall(func, alias) => {
+                self.prepare_function(&func.func, sql);
+                self.prepare_tuple(&func.args, sql);
                 write!(sql, " AS ").unwrap();
                 alias.prepare(sql.as_writer(), self.quote());
             }
@@ -966,7 +972,7 @@ pub trait QueryBuilder: QuotedBuilder + EscapeBuilder + TableRefBuilder {
             #[cfg(feature = "with-mac_address")]
             Value::MacAddress(None) => write!(s, "NULL").unwrap(),
             #[cfg(feature = "postgres-array")]
-            Value::Array(None) => write!(s, "NULL").unwrap(),
+            Value::Array(_, None) => write!(s, "NULL").unwrap(),
             Value::Bool(Some(b)) => write!(s, "{}", if *b { "TRUE" } else { "FALSE" }).unwrap(),
             Value::TinyInt(Some(v)) => write!(s, "{}", v).unwrap(),
             Value::SmallInt(Some(v)) => write!(s, "{}", v).unwrap(),
@@ -1036,7 +1042,7 @@ pub trait QueryBuilder: QuotedBuilder + EscapeBuilder + TableRefBuilder {
             #[cfg(feature = "with-uuid")]
             Value::Uuid(Some(v)) => write!(s, "'{}'", v).unwrap(),
             #[cfg(feature = "postgres-array")]
-            Value::Array(Some(v)) => write!(
+            Value::Array(_, Some(v)) => write!(
                 s,
                 "'{{{}}}'",
                 v.iter()