+++ src/extension/postgres/func.rs
@@ -1,6 +1,6 @@
 //! For calling built-in Postgres SQL functions.
 
-use crate::{expr::*, func::Function};
+use crate::{expr::*, func::*};
 
 /// Functions
 #[derive(Debug, Clone)]
@@ -44,16 +44,17 @@ impl PgFunc {
     ///     r#"SELECT TO_TSQUERY('a & b')"#
     /// );
     /// ```
-    pub fn to_tsquery<T>(expr: T, regconfig: Option<u32>) -> SimpleExpr
+    pub fn to_tsquery<T>(expr: T, regconfig: Option<u32>) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
         match regconfig {
             Some(config) => {
                 let config = SimpleExpr::Value(config.into());
-                Expr::func(Function::PgFunction(PgFunction::ToTsquery)).args([config, expr.into()])
+                FunctionCall::new(Function::PgFunction(PgFunction::ToTsquery))
+                    .args([config, expr.into()])
             }
-            None => Expr::func(Function::PgFunction(PgFunction::ToTsquery)).arg(expr),
+            None => FunctionCall::new(Function::PgFunction(PgFunction::ToTsquery)).arg(expr),
         }
     }
 
@@ -76,16 +77,17 @@ impl PgFunc {
     ///     r#"SELECT TO_TSVECTOR('a b')"#
     /// );
     /// ```
-    pub fn to_tsvector<T>(expr: T, regconfig: Option<u32>) -> SimpleExpr
+    pub fn to_tsvector<T>(expr: T, regconfig: Option<u32>) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
         match regconfig {
             Some(config) => {
                 let config = SimpleExpr::Value(config.into());
-                Expr::func(Function::PgFunction(PgFunction::ToTsvector)).args([config, expr.into()])
+                FunctionCall::new(Function::PgFunction(PgFunction::ToTsvector))
+                    .args([config, expr.into()])
             }
-            None => Expr::func(Function::PgFunction(PgFunction::ToTsvector)).arg(expr),
+            None => FunctionCall::new(Function::PgFunction(PgFunction::ToTsvector)).arg(expr),
         }
     }
 
@@ -108,17 +110,17 @@ impl PgFunc {
     ///     r#"SELECT PHRASETO_TSQUERY('a b')"#
     /// );
     /// ```
-    pub fn phraseto_tsquery<T>(expr: T, regconfig: Option<u32>) -> SimpleExpr
+    pub fn phraseto_tsquery<T>(expr: T, regconfig: Option<u32>) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
         match regconfig {
             Some(config) => {
                 let config = SimpleExpr::Value(config.into());
-                Expr::func(Function::PgFunction(PgFunction::PhrasetoTsquery))
+                FunctionCall::new(Function::PgFunction(PgFunction::PhrasetoTsquery))
                     .args([config, expr.into()])
             }
-            None => Expr::func(Function::PgFunction(PgFunction::PhrasetoTsquery)).arg(expr),
+            None => FunctionCall::new(Function::PgFunction(PgFunction::PhrasetoTsquery)).arg(expr),
         }
     }
 
@@ -141,17 +143,17 @@ impl PgFunc {
     ///     r#"SELECT PLAINTO_TSQUERY('a b')"#
     /// );
     /// ```
-    pub fn plainto_tsquery<T>(expr: T, regconfig: Option<u32>) -> SimpleExpr
+    pub fn plainto_tsquery<T>(expr: T, regconfig: Option<u32>) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
         match regconfig {
             Some(config) => {
                 let config = SimpleExpr::Value(config.into());
-                Expr::func(Function::PgFunction(PgFunction::PlaintoTsquery))
+                FunctionCall::new(Function::PgFunction(PgFunction::PlaintoTsquery))
                     .args([config, expr.into()])
             }
-            None => Expr::func(Function::PgFunction(PgFunction::PlaintoTsquery)).arg(expr),
+            None => FunctionCall::new(Function::PgFunction(PgFunction::PlaintoTsquery)).arg(expr),
         }
     }
 
@@ -174,17 +176,19 @@ impl PgFunc {
     ///     r#"SELECT WEBSEARCH_TO_TSQUERY('a b')"#
     /// );
     /// ```
-    pub fn websearch_to_tsquery<T>(expr: T, regconfig: Option<u32>) -> SimpleExpr
+    pub fn websearch_to_tsquery<T>(expr: T, regconfig: Option<u32>) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
         match regconfig {
             Some(config) => {
                 let config = SimpleExpr::Value(config.into());
-                Expr::func(Function::PgFunction(PgFunction::WebsearchToTsquery))
+                FunctionCall::new(Function::PgFunction(PgFunction::WebsearchToTsquery))
                     .args([config, expr.into()])
             }
-            None => Expr::func(Function::PgFunction(PgFunction::WebsearchToTsquery)).arg(expr),
+            None => {
+                FunctionCall::new(Function::PgFunction(PgFunction::WebsearchToTsquery)).arg(expr)
+            }
         }
     }
 
@@ -204,11 +208,12 @@ impl PgFunc {
     ///     r#"SELECT TS_RANK('a b', 'a&b')"#
     /// );
     /// ```
-    pub fn ts_rank<T>(vector: T, query: T) -> SimpleExpr
+    pub fn ts_rank<T>(vector: T, query: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::PgFunction(PgFunction::TsRank)).args([vector, query])
+        FunctionCall::new(Function::PgFunction(PgFunction::TsRank))
+            .args([vector.into(), query.into()])
     }
 
     /// Call `TS_RANK_CD` function. Postgres only.
@@ -227,11 +232,12 @@ impl PgFunc {
     ///     r#"SELECT TS_RANK_CD('a b', 'a&b')"#
     /// );
     /// ```
-    pub fn ts_rank_cd<T>(vector: T, query: T) -> SimpleExpr
+    pub fn ts_rank_cd<T>(vector: T, query: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::PgFunction(PgFunction::TsRankCd)).args([vector, query])
+        FunctionCall::new(Function::PgFunction(PgFunction::TsRankCd))
+            .args([vector.into(), query.into()])
     }
 
     /// Call `ANY` function. Postgres only.
@@ -251,11 +257,11 @@ impl PgFunc {
     /// );
     /// ```
     #[cfg(feature = "postgres-array")]
-    pub fn any<T>(expr: T) -> SimpleExpr
+    pub fn any<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::PgFunction(PgFunction::Any)).arg(expr)
+        FunctionCall::new(Function::PgFunction(PgFunction::Any)).arg(expr)
     }
 
     /// Call `SOME` function. Postgres only.
@@ -275,11 +281,11 @@ impl PgFunc {
     /// );
     /// ```
     #[cfg(feature = "postgres-array")]
-    pub fn some<T>(expr: T) -> SimpleExpr
+    pub fn some<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::PgFunction(PgFunction::Some)).arg(expr)
+        FunctionCall::new(Function::PgFunction(PgFunction::Some)).arg(expr)
     }
 
     /// Call `ALL` function. Postgres only.
@@ -299,10 +305,10 @@ impl PgFunc {
     /// );
     /// ```
     #[cfg(feature = "postgres-array")]
-    pub fn all<T>(expr: T) -> SimpleExpr
+    pub fn all<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::PgFunction(PgFunction::All)).arg(expr)
+        FunctionCall::new(Function::PgFunction(PgFunction::All)).arg(expr)
     }
 }