+++ src/func.rs
@@ -1,6 +1,6 @@
 //! For calling built-in SQL functions.
 
-use crate::{expr::*, types::*, Value};
+use crate::{expr::*, types::*};
 
 #[cfg(feature = "backend-postgres")]
 pub use crate::extension::postgres::{PgFunc, PgFunction};
@@ -26,6 +26,38 @@ pub enum Function {
     PgFunction(PgFunction),
 }
 
+/// Function call.
+#[derive(Debug, Clone)]
+pub struct FunctionCall {
+    pub(crate) func: Function,
+    pub(crate) args: Vec<SimpleExpr>,
+}
+
+impl FunctionCall {
+    pub(crate) fn new(func: Function) -> Self {
+        Self {
+            func,
+            args: Vec::new(),
+        }
+    }
+
+    pub fn arg<T>(mut self, arg: T) -> Self
+    where
+        T: Into<SimpleExpr>,
+    {
+        self.args = vec![arg.into()];
+        self
+    }
+
+    pub fn args<I>(mut self, args: I) -> Self
+    where
+        I: IntoIterator<Item = SimpleExpr>,
+    {
+        self.args = args.into_iter().collect();
+        self
+    }
+}
+
 /// Function call helper.
 #[derive(Debug, Clone)]
 pub struct Func;
@@ -47,7 +79,7 @@ impl Func {
     /// }
     ///
     /// let query = Query::select()
-    ///     .expr(Func::cust(MyFunction).args([Expr::val("hello")]))
+    ///     .expr(Func::cust(MyFunction).arg(Expr::val("hello")))
     ///     .to_owned();
     ///
     /// assert_eq!(
@@ -63,11 +95,11 @@ impl Func {
     ///     r#"SELECT MY_FUNCTION('hello')"#
     /// );
     /// ```
-    pub fn cust<T>(func: T) -> Expr
+    pub fn cust<T>(func: T) -> FunctionCall
     where
         T: IntoIden,
     {
-        Expr::func(Function::Custom(func.into_iden()))
+        FunctionCall::new(Function::Custom(func.into_iden()))
     }
 
     /// Call `MAX` function.
@@ -95,11 +127,11 @@ impl Func {
     ///     r#"SELECT MAX("character"."size_w") FROM "character""#
     /// );
     /// ```
-    pub fn max<T>(expr: T) -> SimpleExpr
+    pub fn max<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Max).arg(expr)
+        FunctionCall::new(Function::Max).arg(expr)
     }
 
     /// Call `MIN` function.
@@ -127,11 +159,11 @@ impl Func {
     ///     r#"SELECT MIN("character"."size_h") FROM "character""#
     /// );
     /// ```
-    pub fn min<T>(expr: T) -> SimpleExpr
+    pub fn min<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Min).arg(expr)
+        FunctionCall::new(Function::Min).arg(expr)
     }
 
     /// Call `SUM` function.
@@ -159,11 +191,11 @@ impl Func {
     ///     r#"SELECT SUM("character"."size_h") FROM "character""#
     /// );
     /// ```
-    pub fn sum<T>(expr: T) -> SimpleExpr
+    pub fn sum<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Sum).arg(expr)
+        FunctionCall::new(Function::Sum).arg(expr)
     }
 
     /// Call `AVG` function.
@@ -191,11 +223,11 @@ impl Func {
     ///     r#"SELECT AVG("character"."size_h") FROM "character""#
     /// );
     /// ```
-    pub fn avg<T>(expr: T) -> SimpleExpr
+    pub fn avg<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Avg).arg(expr)
+        FunctionCall::new(Function::Avg).arg(expr)
     }
 
     /// Call `ABS` function.
@@ -223,11 +255,11 @@ impl Func {
     ///     r#"SELECT ABS("character"."size_h") FROM "character""#
     /// );
     /// ```
-    pub fn abs<T>(expr: T) -> SimpleExpr
+    pub fn abs<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Abs).arg(expr)
+        FunctionCall::new(Function::Abs).arg(expr)
     }
 
     /// Call `COUNT` function.
@@ -255,11 +287,11 @@ impl Func {
     ///     r#"SELECT COUNT("character"."id") FROM "character""#
     /// );
     /// ```
-    pub fn count<T>(expr: T) -> SimpleExpr
+    pub fn count<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Count).arg(expr)
+        FunctionCall::new(Function::Count).arg(expr)
     }
 
     /// Call `CHAR_LENGTH` function.
@@ -287,11 +319,11 @@ impl Func {
     ///     r#"SELECT LENGTH("character"."character") FROM "character""#
     /// );
     /// ```
-    pub fn char_length<T>(expr: T) -> SimpleExpr
+    pub fn char_length<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::CharLength).arg(expr)
+        FunctionCall::new(Function::CharLength).arg(expr)
     }
 
     /// Call `IF NULL` function.
@@ -322,12 +354,12 @@ impl Func {
     ///     r#"SELECT IFNULL("size_w", "size_h") FROM "character""#
     /// );
     /// ```
-    pub fn if_null<A, B>(a: A, b: B) -> SimpleExpr
+    pub fn if_null<A, B>(a: A, b: B) -> FunctionCall
     where
         A: Into<SimpleExpr>,
         B: Into<SimpleExpr>,
     {
-        Expr::func(Function::IfNull).args([a.into(), b.into()])
+        FunctionCall::new(Function::IfNull).args([a.into(), b.into()])
     }
 
     /// Call `CAST` function with a custom type.
@@ -354,12 +386,13 @@ impl Func {
     ///     r#"SELECT CAST('hello' AS MyType)"#
     /// );
     /// ```
-    pub fn cast_as<V, I>(value: V, iden: I) -> SimpleExpr
+    pub fn cast_as<V, I>(expr: V, iden: I) -> FunctionCall
     where
-        V: Into<Value>,
+        V: Into<SimpleExpr>,
         I: IntoIden,
     {
-        Expr::func(Function::Cast).arg(Expr::val(value.into()).bin_oper(
+        let expr: SimpleExpr = expr.into();
+        FunctionCall::new(Function::Cast).arg(expr.binary(
             BinOper::As,
             Expr::cust(iden.into_iden().to_string().as_str()),
         ))
@@ -374,9 +407,9 @@ impl Func {
     ///
     /// let query = Query::select()
     ///     .expr(Func::coalesce([
-    ///         Expr::col(Char::SizeW),
-    ///         Expr::col(Char::SizeH),
-    ///         Expr::val(12),
+    ///         Expr::col(Char::SizeW).into(),
+    ///         Expr::col(Char::SizeH).into(),
+    ///         Expr::val(12).into(),
     ///     ]))
     ///     .from(Char::Table)
     ///     .to_owned();
@@ -394,12 +427,11 @@ impl Func {
     ///     r#"SELECT COALESCE("size_w", "size_h", 12) FROM "character""#
     /// );
     /// ```
-    pub fn coalesce<I, T>(args: I) -> SimpleExpr
+    pub fn coalesce<I>(args: I) -> FunctionCall
     where
-        T: Into<SimpleExpr>,
-        I: IntoIterator<Item = T>,
+        I: IntoIterator<Item = SimpleExpr>,
     {
-        Expr::func(Function::Coalesce).args(args)
+        FunctionCall::new(Function::Coalesce).args(args)
     }
 
     /// Call `LOWER` function.
@@ -407,7 +439,6 @@ impl Func {
     /// # Examples
     ///
     /// ```
-    /// use sea_query::tests_cfg::Character::Character;
     /// use sea_query::{tests_cfg::*, *};
     ///
     /// let query = Query::select()
@@ -428,11 +459,11 @@ impl Func {
     ///     r#"SELECT LOWER("character") FROM "character""#
     /// );
     /// ```
-    pub fn lower<T>(expr: T) -> SimpleExpr
+    pub fn lower<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Lower).arg(expr)
+        FunctionCall::new(Function::Lower).arg(expr)
     }
 
     /// Call `UPPER` function.
@@ -440,7 +471,6 @@ impl Func {
     /// # Examples
     ///
     /// ```
-    /// use sea_query::tests_cfg::Character::Character;
     /// use sea_query::{tests_cfg::*, *};
     ///
     /// let query = Query::select()
@@ -461,11 +491,11 @@ impl Func {
     ///     r#"SELECT UPPER("character") FROM "character""#
     /// );
     /// ```
-    pub fn upper<T>(expr: T) -> SimpleExpr
+    pub fn upper<T>(expr: T) -> FunctionCall
     where
         T: Into<SimpleExpr>,
     {
-        Expr::func(Function::Upper).arg(expr)
+        FunctionCall::new(Function::Upper).arg(expr)
     }
 
     /// Call `RANDOM` function.
@@ -484,7 +514,7 @@ impl Func {
     ///
     /// assert_eq!(query.to_string(SqliteQueryBuilder), r#"SELECT RANDOM()"#);
     /// ```
-    pub fn random() -> SimpleExpr {
-        Expr::func(Function::Random).into()
+    pub fn random() -> FunctionCall {
+        FunctionCall::new(Function::Random)
     }
 }